Compression & Decompression functions work. The train, compress & decompress modes work as intended.

+++ modules/models.py
@@ -6,15 +6,15 @@ class george_SAE(nn.Module):
     def __init__(self, n_features, z_dim):
         super(george_SAE, self).__init__()
         # encoder
-        self.en1 = nn.Linear(n_features, 200)
-        self.en2 = nn.Linear(200, 100)
-        self.en3 = nn.Linear(100, 50)
-        self.en4 = nn.Linear(50, z_dim)
+        self.en1 = nn.Linear(n_features, 200,dtype=torch.float64)
+        self.en2 = nn.Linear(200, 100,dtype=torch.float64)
+        self.en3 = nn.Linear(100, 50,dtype=torch.float64)
+        self.en4 = nn.Linear(50, z_dim,dtype=torch.float64)
         # decoder
-        self.de1 = nn.Linear(z_dim, 50)
-        self.de2 = nn.Linear(50, 100)
-        self.de3 = nn.Linear(100, 200)
-        self.de4 = nn.Linear(200, n_features)
+        self.de1 = nn.Linear(z_dim, 50,dtype=torch.float64)
+        self.de2 = nn.Linear(50, 100,dtype=torch.float64)
+        self.de3 = nn.Linear(100, 200,dtype=torch.float64)
+        self.de4 = nn.Linear(200, n_features,dtype=torch.float64)
 
         self.n_features = n_features
         self.z_dim = z_dim
